+++ src/main/java/lotto/LottoPlayer.java
@@ -1,6 +1,11 @@
 package lotto;
 
 import lotto.domain.*;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.lotto.Lottos;
+import lotto.domain.rank.Ranks;
+import lotto.domain.statistics.WinningStatistics;
 import lotto.view.InputView;
 import lotto.view.OutputView;
