+++ src/main/java/lotto/domain/LotteryShop.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.domain.lotto.AutoLottoGenerator;
+import lotto.domain.lotto.Lottos;
+
 public class LotteryShop {
     public static final int LOTTO_PRICE = 1000;
