+++ src/main/java/lotto/domain/lotto/AutoLottoGenerator.java
@@ -1,4 +1,8 @@
-package lotto.domain;
+package lotto.domain.lotto;
+
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.lotto.Lottos;
 
 import java.util.ArrayList;
 import java.util.Collections;
