+++ src/main/java/lotto/domain/rank/Rank.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.rank;
+
+import lotto.domain.lotto.LottoMatchCount;
 
 import java.util.Arrays;
 import java.util.List;
