+++ src/main/java/lotto/domain/statistics/Count.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.statistics;
 
 import java.util.Objects;
