+++ src/test/java/lotto/domain/lotto/AutoLottoGeneratorTest.java
@@ -1,9 +1,8 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.HashSet;
